made changes as per pr and good to merge
